More progress.
Implemented (very primitive) parsing of source files to extract a hierarchy of
packages and classes. Wired up rebuilding of index files, and added fqName
thereto to simplify life for the import processor (it no longer has to read the
file to extract the fqName). This allows us to handle inner classes properly as
well.