Revamped index/cache file management.
Now instead of having a .findclass.cache file associated with each
.findclass.path, we instead maintain a cache file for each source path that
occurs in any .findclass.path in any project. So if two projects depend on the
same source directory, they'll share an index file.

This also opens the door to having indices based not on source directories, but
based on jar files (containing sources or classes).